Write script to generate source listing manually
Although automatic source listing generation is apparently not as
straightforward as I hoped, it can be made very very fast using a
combination of vim commands and a python script (specifically, it takes
me 8 keystrokes (die;.!<up><cr>) to replace the contents of source.tex).

With this, we can consider the source code as good as listed.