Modify converter.c to use a table of Morse codes
Converting 'every single character' being fed to the program would take
a lot of time and memory overhead, and could crash the system with
larger inputs. Rather than doing that, we can instead make a static
table of known Morse codes and modify the conversion function to return
the index into that table for a character.