Remove bare metal RPi3 utilities
As explained in a previous commit, bare-metal utilities won't work
without triggering a segfault while the OS is active. Due to this, none
of the utilities are being used in the program, meaning they can be
safely eliminated (except for `rpi3/blink.h`, which was modified to use
wiringPi for the core functionality).