Update get_index function in converter.c
This commit replaces the giant switch statement with three if statements
for readability and efficiency. Now code folding in the editor is no
longer necessary to reasonably navigate the file!

Since converter.c does not really need the get_index in the public
interface (since nothing will use it), the prototype has been removed
from the header file and the function was changed to static.